Create Merging_Overlapping_Intervals.cpp
The idea is to sort the intervals in increasing order of their starting time. Then create an empty stack and for each interval,
If the stack is empty or the top interval in the stack does not overlap with the current interval, push it into the stack.
If the top interval of the stack overlaps with the current interval, merge both intervals by updating the end of the top interval at the ending of the current interval.
Finally, print all non-overlapping intervals present in the stack.